Split proof-of-concept code into its own header.
